pyproject.toml
[project]
name = "geological-marker"
version = "0.1.0"
description = "Make geological map with AIST map"
authors = [
    {name = "Satoshi Matsuno",email = "51268793+MSrakugo@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit (>=1.47.1,<2.0.0)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "folium (>=0.20.0,<0.21.0)",
    "streamlit-folium (>=0.25.1,<0.26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
